+++ RMPhoneFormat/RMViewController.m
@@ -109,8 +109,8 @@
         }
         
         // Now let's find the position, in the newly formatted string, of the same number of non-formatting characters.
-        int pos = [phone length];
-        int cnt2 = 0;
+        NSUInteger pos = [phone length];
+        NSUInteger cnt2 = 0;
         for (NSUInteger i = 0; i < [phone length]; i++) {
             if ([_phoneChars characterIsMember:[phone characterAtIndex:i]]) {
                 cnt2++;
